Merging 1.3.1 to master for release

+++ fieldcapture-hubs-plugin/grails-app/services/au/org/ala/fieldcapture/DocumentService.groovy
@@ -1,6 +1,7 @@
 package au.org.ala.fieldcapture
 
 import grails.converters.JSON
+import static org.apache.http.HttpStatus.*;
 
 /**
  * Proxies to the ecodata DocumentController/DocumentService.
@@ -15,9 +16,24 @@ class DocumentService {
         return webService.getJson(url)
     }
 
+    def delete(String id) {
+        def url = "${grailsApplication.config.ecodata.baseUrl}document/${id}"
+        return webService.doDelete(url)
+    }
+
+
     def createTextDocument(doc, content) {
         doc.content = content
         updateDocument(doc)
+    }
+
+    def findAllHelpResources() {
+        def url = "${grailsApplication.config.ecodata.baseUrl}document/search"
+        def result = webService.doPost(url, [role:'helpResource'])
+        if (result.statusCode == SC_OK) {
+            return result.resp.documents
+        }
+        return []
     }
 
     def updateDocument(doc) {
@@ -26,12 +42,17 @@ class DocumentService {
         return webService.doPost(url, doc)
     }
 
-    def createDocument(doc, contentType, inputStream) {
-
+    def updateDocument(doc, contentType, inputStream) {
         def url = grailsApplication.config.ecodata.baseUrl + "document"
-
+        if (doc.documentId) {
+            url+="/"+doc.documentId
+        }
         def params = [document:doc as JSON]
         return webService.postMultipart(url, params, inputStream, contentType, doc.filename)
+    }
+
+    def createDocument(doc, contentType, inputStream) {
+        updateDocument(doc, contentType, inputStream)
     }
 
     def getDocumentsForSite(id) {
@@ -49,14 +70,20 @@ class DocumentService {
         def result
         if (!document.documentId) {
             document.remove('url')
-            def file = new File(grailsApplication.config.upload.images.path, document.filename)
-            // Create a new document, supplying the file that was uploaded to the ImageController.
-            result = createDocument(document, document.contentType, new FileInputStream(file))
-            if (!result.error) {
-                file.delete()
+            File file = new File(grailsApplication.config.upload.images.path, document.filename)
+            if (file.exists()) {
+                // Create a new document, supplying the file that was uploaded to the ImageController.
+                result = createDocument(document, document.contentType, new FileInputStream(file))
+                if (!result.error) {
+                    file.delete()
+                }
+            }
+            else {
+                result = updateDocument(document)
             }
         }
         else {
+
             // Just update the document.
             result = updateDocument(document)
         }
@@ -68,5 +95,14 @@ class DocumentService {
         link.type = "link"
         link.externalUrl = link.remove('url')
         updateDocument(link)
+    }
+
+    def Map search(Map params) {
+        def url = "${grailsApplication.config.ecodata.baseUrl}document/search"
+        def resp = webService.doPost(url, params)
+        if (resp && !resp.error) {
+            return resp.resp
+        }
+        return resp
     }
 }
